+++ blelitelib/src/main/java/com/afunx/ble/blelitelib/operation/BleWriteCharacteristicNoResponse20Operation.java
@@ -67,7 +67,7 @@ public class BleWriteCharacteristicNoResponse20Operation extends BleOperationAbs
     }
 
     @Override
-    public long getOperatcionCode() {
+    public int getOperatcionCode() {
         return BLE_WRITE_CHARACTERISTIC_NO_RESPONSE_20;
     }
 }
